% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustHhVehicleMpgMpkwh.R
\name{AdjustHhVehicleMpgMpkwh}
\alias{AdjustHhVehicleMpgMpkwh}
\title{Adjust energy efficiency and emissions rates of vehicles used by households.}
\usage{
AdjustHhVehicleMpgMpkwh(L, M)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}

\item{M}{A list the module functions of modules called by this module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AdjustHhVehicleMpgMpkwh} adjusts the energy efficiency and emissions of
vehicles used by households.
}
\details{
This function adjusts the energy efficiency and emissions of
vehicles used by households as a consequence of roadway congestion, speed
smoothing, and ecodriving.
}
