% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustHhVehicleMpgMpkwh.R
\docType{data}
\name{AdjustHhVehicleMpgMpkwhSpecifications}
\alias{AdjustHhVehicleMpgMpkwhSpecifications}
\title{Specifications list for AdjustHhVehicleMpgMpkwh module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AdjustHhVehicleMpgMpkwh.R script.
}
\usage{
AdjustHhVehicleMpgMpkwhSpecifications
}
\description{
A list containing specifications for the AdjustHhVehicleMpgMpkwh module.
}
\keyword{datasets}
