% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadDvmt.R
\name{assignHhUrbanDvmtProp}
\alias{assignHhUrbanDvmtProp}
\title{Assign household DVMT proportion occurring on urban roads}
\usage{
assignHhUrbanDvmtProp(Hh_ls, HhUrbanRoadDvmt_Ma)
}
\arguments{
\item{Hh_ls}{A list containing the characteristics of households used to
allocate the proportion of household DVMT occurring on urban area roads. The
list must include the marea name for each household}

\item{HhUrbanRoadDvmt_Ma}{A named numeric vector identifying the total amount
of household DVMT in each marea that occurs on urban area roads.}
}
\value{
A numeric vector identifying the proportion of the household DVMT
occurring on marea urban roadways.
}
\description{
\code{assignHhUrbanDvmtProp} assigns a proportion of the DVMT of each
marea household to urban roads in the marea such that the total is consistent
with the total household DVMT assigned to urban roadways.
}
\details{
This function is called by the CalculateRoadDvmt to assign the proportion of
each marea household's DVMT that occurs on urban roadways in the marea. The
proportions are calculated to be consistent with the control total calculated
for the marea.
}
