% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceRoadCostsAndRevenues.R
\name{BalanceRoadCostsAndRevenues}
\alias{BalanceRoadCostsAndRevenues}
\title{Calculate household VMT tax to balance road costs and revenues.}
\usage{
BalanceRoadCostsAndRevenues(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{BalanceRoadCostsAndRevenues} calculates an extra VMT tax needed to
balance road costs and road revenues.
}
\details{
This function calculates roadway costs that are attributable to households.
It compares those costs with the road tax revenues (fuel tax, equivalent tax
paid by plug-in vehicles, VMT tax, and congestion charges) and identifies
whether the road tax revenues are sufficient to cover costs. If not, it
calculates an extra VMT tax rate to cover the costs.
}
