% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceRoadCostsAndRevenues.R
\docType{data}
\name{BalanceRoadCostsAndRevenuesSpecifications}
\alias{BalanceRoadCostsAndRevenuesSpecifications}
\title{Specifications list for BalanceRoadCostsAndRevenues module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
BalanceRoadCostsAndRevenues.R script.
}
\usage{
BalanceRoadCostsAndRevenuesSpecifications
}
\description{
A list containing specifications for the BalanceRoadCostsAndRevenues
module.
}
\keyword{datasets}
