% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\docType{data}
\name{BaseSpeeds_df}
\alias{BaseSpeeds_df}
\title{Base arterial and freeway speeds by congestion level}
\format{A data frame with 5 rows and 4 columns. Row names are congestion
level names(None, Mod, Hvy, Sev, Ext):
\describe{
  \item{Fwy}{average freeway speed in miles per hours by congestion level
  considering the effects of recurring and incident-related congestion}
  \item{Art}{average arterial speed in miles per hours by congestion level
  considering the effects of recurring and incident-related congestion}
  \item{Fwy_Rcr}{average freeway speed in miles per hours by congestion level
  considering only the effects of recurring congestion}
  \item{Art_Rcr}{average arterial speed in miles per hours by congestion
  level considering the effects of recurring and incident-related congestion}
  }}
\source{
CalculateSpeeds.R script.
}
\usage{
BaseSpeeds_df
}
\description{
Base speeds on freeways and arterials by congestion level
}
\keyword{datasets}
