% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudgetHouseholdDvmt.R
\docType{data}
\name{BudgetHouseholdDvmtSpecifications}
\alias{BudgetHouseholdDvmtSpecifications}
\title{Specifications list for BudgetHouseholdDvmt module}
\format{A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
 \item{Call}{modules that are called by the module}
}}
\source{
BudgetHouseholdDvmt.R script.
}
\usage{
BudgetHouseholdDvmtSpecifications
}
\description{
A list containing specifications for the BudgetHouseholdDvmt module.
}
\keyword{datasets}
