% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudgetHouseholdDvmt.R
\docType{data}
\name{BudgetModel_ls}
\alias{BudgetModel_ls}
\title{Household vehicle operating cost budget model}
\format{A list:
\describe{
  \item{MaxOpProp_SS}{a smooth spline model to predict the maximum operating cost proportion of income}
  \item{MaxValue}{a scalar value that is the maximum proportion}
}}
\source{
BudgetHouseholdDvmt.R script.
}
\usage{
BudgetModel_ls
}
\description{
A smooth spline model to predict the maximum operating cost budget as a function of income and a maximum value to use for very low income households.
}
\keyword{datasets}
