% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBaseRoadDvmt.R
\docType{data}
\name{CalculateBaseRoadDvmtSpecifications}
\alias{CalculateBaseRoadDvmtSpecifications}
\title{Specifications list for CalculateBaseRoadDvmt module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateBaseRoadDvmt.R script.
}
\usage{
CalculateBaseRoadDvmtSpecifications
}
\description{
A list containing specifications for the CalculateBaseRoadDvmt module.
}
\keyword{datasets}
