% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateComEnergyAndEmissions.R
\name{CalculateComEnergyAndEmissions}
\alias{CalculateComEnergyAndEmissions}
\title{Calculate energy and emissions of commercial vehicle travel.}
\usage{
CalculateComEnergyAndEmissions(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateComEnergyAndEmissions} calculates the hydrocarbon and
electrical energy consumption of the travel of commercial service vehicles
and heavy trucks.
}
\details{
This function calculates the hydrocarbon and electrical energy consumption of
the travel of commercial service vehicles and heavy trucks and the carbon
emissions produced.
}
