% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\name{calculateCongestion}
\alias{calculateCongestion}
\title{Calculate DVMT or DVHT by congestion level.}
\usage{
calculateCongestion(RoadType, LaneMi, DVMT)
}
\arguments{
\item{RoadType}{a string having the value 'Fwy' or 'Art' depending on whether
congestion is to be calculated for freeways (Fwy) or arterial roadways (Art).}

\item{LaneMi}{a number identifying the number of lane-miles for the roadway
type.}

\item{DVMT}{a number identifying the daily vehicle miles of travel on the
roadway type.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateCongestion} splits input DVMT into the amounts in
each of 5 congestion levels:None, Mod, Hvy, Sev, Ext.
}
\details{
This function splits input DVMT into amounts by congestion level
(none, moderate, heavy, severe, and extreme) as a function of the ratio of
DVMT to lane-miles. The lookup tables in CongestedProportions_ls are used
to lookup proportions by congestion level. The input DVMT is
split with these proportions and the split values are returned.
}
