% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFutureRoadDvmt.R
\docType{data}
\name{CalculateFutureRoadDvmtSpecifications}
\alias{CalculateFutureRoadDvmtSpecifications}
\title{Specifications list for CalculateFutureRoadDvmt module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateFutureRoadDvmt.R script.
}
\usage{
CalculateFutureRoadDvmtSpecifications
}
\description{
A list containing specifications for the CalculateFutureRoadDvmt module.
}
\keyword{datasets}
