% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMpgMpkwhAdjustments.R
\docType{data}
\name{CalculateMpgMpkwhAdjustmentsSpecifications}
\alias{CalculateMpgMpkwhAdjustmentsSpecifications}
\title{Specifications list for CalculateMpgMpkwhAdjustments module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateMpgMpkwhAdjustments.R script.
}
\usage{
CalculateMpgMpkwhAdjustmentsSpecifications
}
\description{
A list containing specifications for the CalculateCarbonIntensity module.
}
\keyword{datasets}
