% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePtranEnergyAndEmissions.R
\name{CalculatePtranEnergyAndEmissions}
\alias{CalculatePtranEnergyAndEmissions}
\title{Calculate energy and emissions of public transit vehicle travel.}
\usage{
CalculatePtranEnergyAndEmissions(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculatePtranEnergyAndEmissions} calculates the hydrocarbon and
electrical energy consumption of the travel of public transit vehicles.
}
\details{
This function calculates the hydrocarbon and electrical energy consumption of
the travel of public transit vehicles and the carbon emissions produced.
}
