% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePtranEnergyAndEmissions.R
\docType{data}
\name{CalculatePtranEnergyAndEmissionsSpecifications}
\alias{CalculatePtranEnergyAndEmissionsSpecifications}
\title{Specifications list for CalculatePtranEnergyAndEmissions module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculatePtranEnergyAndEmissions.R script.
}
\usage{
CalculatePtranEnergyAndEmissionsSpecifications
}
\description{
A list containing specifications for the CalculatePtranEnergyAndEmissions
module.
}
\keyword{datasets}
