% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\name{CalculateRoadPerformance}
\alias{CalculateRoadPerformance}
\title{Calculate road performance measures}
\usage{
CalculateRoadPerformance(L)
}
\arguments{
\item{L}{A list containing data defined by the module specification.}
}
\value{
A list containing data produced by the function consistent with the
module specifications.
}
\description{
\code{CalculateRoadPerformance} calculates freeway and arterial daily vehicle
miles of travel (DVMT) by congestion level, freeway and arterial speeds and
delays by congestion level, and the split of light-duty vehicle DVMT between
freeways and arterials.
}
\details{
This function splits freeway DVMT by congestion level based on the ratio of
DVMT to lane miles using a lookup table. Arterial DVMT is split into
congestion levels in the same manner. Freeway speeds and arterial speeds
by congestion level are calculated given the user-specified deployments of
operations measures (e.g. ramp metering, incident management, signal
coordination, access management). The function likewise calculates delay rate
(hours per mile) by congestion level. User-specified congestion prices are
converted into equivalent vehicle hours of travel (VHT) based using the
user-defined value-of-time parameter. The sum of actual VHT and these
equivalent VHT are divided by total DVMT to calculate an average equivalent
speed. Light-duty vehicle DVMT is split between freeways and arterials based
on the ratio of freeway and arterial average equivalent speeds and a
parameter that is calculated from the DVMT split model. These calculations
are repeated until an equilibrium value is found (i.e. the DVMT split ratio
changes by less than 0.01% from one iteration to the next).
}
