% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSafetyMeasures.R
\docType{data}
\name{CalculateSafetyMeasuresSpecifications}
\alias{CalculateSafetyMeasuresSpecifications}
\title{Specifications list for CalculateRoadDvmt module}
\format{
A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}
}
\source{
CalculateSafetyMeasures.R script.
}
\usage{
CalculateSafetyMeasuresSpecifications
}
\description{
A list containing specifications for the CalculateRoadDvmt module.
}
\keyword{datasets}
