% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\name{calculateSpeeds}
\alias{calculateSpeeds}
\title{Calculate speeds and delay by road class and congestion level.}
\usage{
calculateSpeeds(OpsDeployment_, OtherOpsEffects_mx = NULL)
}
\arguments{
\item{OpsDeployment_}{a numeric vector identifying the proportional degree of
roadway operations programs (ramp metering, incident management, signal
coordination, accessmanagement, other) where 0 means no deployment and
1 means deployment covering the entire road system.}

\item{OtherOpsEffects_mx}{a numeric matrix of percentage delay reductions
for user-defined freeway and arterial operations programs by congestion
level.}
}
\value{
A list containing two matrices. The first is a matrix of average
speed (miles per hour) by congestion level and road class. The second is a
matrix of delay hours per vehicle mile by congestion level and road class.
}
\description{
\code{calculateSpeeds} Calculate and speeds and delay by road class and
congestion level.
}
\details{
This function calculates roadway speeds and delay by road class and
congestion level in consideration of the effects of reductions in delay due
to the deployment of operations measures.
}
