% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleOperatingCost.R
\name{CalculateVehicleOperatingCost}
\alias{CalculateVehicleOperatingCost}
\title{Calculate vehicle operating costs.}
\usage{
CalculateVehicleOperatingCost(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateVehicleOperatingCost} calculates vehicle operating costs and
determines how household DVMT is split between vehicles.
}
\details{
This function calculates vehicle operating costs, splits household DVMT
between vehicles, and calculates household average vehicle operating cost,
social/environmental impact cost, and road use taxes per mile of household
vehicle travel.
}
