% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleOperatingCost.R
\docType{data}
\name{CalculateVehicleOperatingCostSpecifications}
\alias{CalculateVehicleOperatingCostSpecifications}
\title{Specifications list for CalculateVehicleOperatingCost module}
\format{A list containing 5 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{scenario input data to be loaded into the datastore for this
 module}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
 \item{Call}{list of modules called by the module}
}}
\source{
CalculateVehicleOperatingCost.R script.
}
\usage{
CalculateVehicleOperatingCostSpecifications
}
\description{
A list containing specifications for the CalculateVehicleOperatingCost module.
}
\keyword{datasets}
