% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\docType{data}
\name{CongestedProportions_ls}
\alias{CongestedProportions_ls}
\title{Congestion lookup tables of proportions of DVMT by congestion level
for freeways and arterials by demand to supply ratio (average DVMT per lane-mile)}
\format{A list of matrices. The list has two components, Fwy and Art, that
  are lookup tables for DVMT by congestion level for freeways and arterials
  respectively. Each table is a matrix containing 5 columns corresponding the
  5 congestion levels (None, Mod, Hvy, Sev, Ext). The rows correspond to
  demand to supply ratios with the rownames specifying the ratios. The values
  are proportions by congestion level. The values in each row sum to 1.}
\source{
CalculateCongestion.R script.
}
\usage{
CongestedProportions_ls
}
\description{
Congestion lookup tables of proportions of DVMT by congestion level
for freeways and arterials by demand to supply ratio (average DVMT per lane-mile)
}
\keyword{datasets}
