% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\docType{data}
\name{Delay_df}
\alias{Delay_df}
\title{Freeway and arterial recurring and non-recurring delay by congestion level}
\format{A data frame with 5 rows and 4 columns. Row names are congestion
level names(None, Mod, Hvy, Sev, Ext):
\describe{
  \item{Fwy_Rcr}{freeway recurring delay in hours per mile by congestion
  level}
  \item{Fwy_NonRcr}{freeway non-recurring (incident) delay in hours per mile
  by congestion level}
  \item{Art_Rcr}{arterial recurring delay in hours per mile by congestion
  level}
  \item{Art_NonRcr}{arterial non-recurring (incident) delay in hours per mile
  by congestion level}
  }}
\source{
CalculateSpeeds.R script.
}
\usage{
Delay_df
}
\description{
Freeway and arterial recurring and non-recurring delay by congestion level
}
\keyword{datasets}
