% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\docType{data}
\name{Incident_df}
\alias{Incident_df}
\title{Delay reduction due to incident management}
\format{A data frame with 5 rows and 4 columns. Row names are congestion
level names(None, Mod, Hvy, Sev, Ext):
\describe{
  \item{Fwy_Rcr}{percentage reduction in freeway recurring congestion delay
  by congestion level}
  \item{Fwy_NonRcr}{percentage reduction in freeway non-recurring congestion
  delay by congestion level}
  \item{Art_Rcr}{percentage reduction in arterial recurring congestion delay
  by congestion level}
  \item{Art_NonRcr}{percentage reduction in arterial non-recurring congestion
  delay by congestion level}
  }}
\source{
CalculateSpeeds.R script.
}
\usage{
Incident_df
}
\description{
Percentage reduction in recurring and non-recurring (incident) delay from
incident management at full deployment.
}
\keyword{datasets}
