% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\docType{data}
\name{InitializeSpecifications}
\alias{InitializeSpecifications}
\title{Specifications list for Initialize module}
\format{A list containing 2 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{scenario input data to be loaded into the datastore for this
 module}
}}
\source{
Initialize.R script.
}
\usage{
InitializeSpecifications
}
\description{
A list containing specifications for the Initialize module.
}
\keyword{datasets}
