% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudgetHouseholdDvmt.R
\docType{data}
\name{OpPropModel_ls}
\alias{OpPropModel_ls}
\title{Household vehicle operating cost proportion of income model}
\format{A list:
\describe{
  \item{Type}{a string identifying the model type}
  \item{Formula}{a string specifying the model formula}
  \item{PrepFun}{a function to transform function input (i.e. income)}
  \item{OutFun}{a function to transform model output (i.e. to produce proportion)}
  \item{Summary}{summary statistics of linear model}
}}
\source{
BudgetHouseholdDvmt.R script.
}
\usage{
OpPropModel_ls
}
\description{
A list in the format required by the applyLinearModel function. The list
implements a linear model which predicts the maximum proportion of income
that a household is willing to spend on vehicle operating costs.
}
\keyword{datasets}
