% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDefaultRoadDvmtValues.R
\docType{data}
\name{RoadDvmtModel_ls}
\alias{RoadDvmtModel_ls}
\title{Roadway DVMT models}
\format{A list having the following components:
RuralVmt_StVtRc: array of DVMT on non-urban roads by state, vehicle type and
road class used to calculate average split on non-urban roads;
HvyTrkDvmtPC_St: the ratio of heavy truck DVMT to population by state;
HvyTrkDvmtUrbanProp_St: the proportion of heavy truck DVMT occurring within
urban areas of each state;
UzaHvyTrkDvmtPC_Ua: the ratio of heavy truck DVMT to population by urbanized
area;
UzaLDVDvmtPC_Ua: the ratio of light-duty vehicle DVMT to population by
urbanized area;
UrbanRcProps_StVtRc: the proportional split of each type of DVMT by urban
road class in each state;
UzaRcProps_UaVtRc: the proportional split of each type of DVMT by urban
road class in each urbanized area;
ComSvcDvmtFactor: the factor to calculate light-duty commercial service
vehicle DVMT from household DVMT.}
\source{
CalculateRoadDVMT.R
}
\usage{
RoadDvmtModel_ls
}
\description{
A list of components used to predict the roadway DVMT by vehicle type and
road class in urbanized areas.
}
\keyword{datasets}
