% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\docType{data}
\name{UrbanRuralAveSpeed_ls}
\alias{UrbanRuralAveSpeed_ls}
\title{Urban-rural average household speed ratio}
\format{A list
\describe{
  \item{NhtsTripDistTime_df}{data frame containing urban and rural household
  average vehicle trip distance (miles) and trip time (minutes) by census
  region}
  \item{NhtsSpeed_df}{data frame containing urban and rural household
  average vehicle trip speed (miles/hour) and ratio of rural to urban
  average speed}
  \item{UrbanRuralHouseholdSpeedRatio}{a number that is the average speed
  ratio used in the model to calculate rural speed from urban speed}
}}
\source{
CalculateRoadPerformance.R script.
}
\usage{
UrbanRuralAveSpeed_ls
}
\description{
The ratio of average speed of vehicle trips by households residing in urban
areas (i.e. in urbanized areas) to average speed of vehicle trips by
households residing in rural areas (i.e. not in urbanized areas)
}
\keyword{datasets}
