% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\docType{data}
\name{UzaDvmtNames_ls}
\alias{UzaDvmtNames_ls}
\title{Urbanized area names list}
\format{A list containing 2 components:
\describe{
 \item{Names}{a vector of urbanized area names sorted by state and name}
 \item{Table}{a 3-column data frame of the names to include in documentation}
}}
\source{
Initialize.R script.
}
\usage{
UzaDvmtNames_ls
}
\description{
A list containing the names of urbanized areas included in the profiles of
urbanized area DVMT characteristics.
}
\keyword{datasets}
